Переделала jsp на html

+++ src/main/java/ru/innopolis/stc/repository/CourseRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.innopolis.stc.bean.Course;
 
 @Repository
-public interface CourseRepository extends CrudRepository<Course, Long> {
+public interface CourseRepository extends CrudRepository<Course, Integer> {
 }
